+++ darts/models/forecasting/xgboost.py
@@ -136,6 +136,36 @@ class XGBModel(RegressionModel, _LikelihoodMixin):
             that all target `series` have the same static covariate dimensionality in ``fit()`` and ``predict()``.
         **kwargs
             Additional keyword arguments passed to `xgb.XGBRegressor`.
+
+        Examples
+        --------
+        Deterministic forecasting, using past/future covariates (optional)
+        >>> from darts.datasets import WeatherDataset
+        >>> from darts.models import XGBModel
+        >>> series = WeatherDataset().load()
+        >>> # predicting atmospheric pressure
+        >>> target = series['p (mbar)'][:100]
+        >>> # optionally, use past observed rainfall (pretending to be unknown beyond index 100)
+        >>> past_cov = series['rain (mm)'][:100]
+        >>> # optionally, use future temperatures (pretending this component is a forecast)
+        >>> future_cov = series['T (degC)'][:106]
+        >>> # predict 6 pressure values using the 12 past values of pressure and rainfall, as well as the 6 temperature
+        >>> # values corresponding to the forecasted period
+        >>> model = XGBModel(
+        >>>     lags=12,
+        >>>     lags_past_covariates=12,
+        >>>     lags_future_covariates=[0,1,2,3,4,5],
+        >>>     output_chunk_length=6,
+        >>> )
+        >>> model.fit(target, past_covariates=past_cov, future_covariates=future_cov)
+        >>> pred = model.predict(6)
+        >>> pred.values()
+        array([[1005.9185 ],
+               [1005.8315 ],
+               [1005.7878 ],
+               [1005.72626],
+               [1005.7475 ],
+               [1005.76074]])
         """
         kwargs["random_state"] = random_state  # seed for tree learner
         self.kwargs = kwargs